Generate weight equations more effectively.
Instead of substituting particular functions as arguments of the associator and evaluating it at particular points, equate the coefficients of the differential operator to zero.
For polynomial Poisson structures these coefficients are polynomial, so equate the coefficients of these polynomials to zero.
The more general but less effective alternative is to evaluate at particular points again.