LeibnizGraph: inherit from KontsevichGraph
This has the advantage of easy access to the list of targets (by reference, not by value).
The previous constructor used a copy of the list of targets, and created pointers to them.
The copy was subsequently deleted, and the pointers invalidated (dereference => segfault).
The present approach avoids this problem.