KontsevichGraph::operator>>(): don't normalize()
We want to be able to read KontsevichGraphs without normalizing.
For example this is necessary when reading LeibnizGraphs.
Namely the encoding of a LeibnizGraph contains the numbers of the
Jacobiator vertices, which refer to a particular Kontsevich graph
encoding, which may or may not be normalized.

In the other places where operator>> was used before, an explicit
call to normalize() was added, to retain the same functionality.
Namely, the graphs in KontsevichGraphSums and KontsevichGraphSeries
are normalized upon reading from input.